requirements.txt
autopep8==2.0.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
environs==9.5.0
etcd3==0.12.0
filelock==3.12.3
Flask==2.3.3
fsspec==2023.9.2
grpcio==1.58.0
grpcio-tools==1.58.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kafka-python==2.0.2
MarkupSafe==2.1.3
marshmallow==3.20.1
minio==7.1.17
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.0.0
protobuf==4.24.4
pycodestyle==2.11.1
pymilvus==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-etcd==0.4.5
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
ruff==0.0.292
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.17.0
